+++ packages/data-validation/src/types/enum/update-rule-type.ts
@@ -18,4 +18,5 @@ export enum UpdateRuleType {
     SETTING,
     RANGE,
     OPTIONS,
+    ALL,
 }
